requirements.txt
django == 4.0.5
pillow
django-admin-interface
django-rest-framework
social-auth-app-django
django-allauth